feat udistrital/arka_cliente#714: Bodega de consumo
- Refactor consulta de elementos asociados a una solicitud de elementos de bodega de consumo.
